endret sånn at temperatur vises i int

+++ app/src/main/java/com/example/in2000_prosjekt/ui/components/Component.kt
@@ -560,7 +560,7 @@ fun DeleteAllButton(viewModel: FavoriteViewModel){
 }
 
 @Composable
-fun Illustrasjon(height : Int?, temp : Float, vind : Float, weatherHigh : Float, weatherMid : Float, weatherLow : Float){
+fun Illustrasjon(height: Int?, temp: Int, vind: Float, weatherHigh: Float, weatherMid: Float, weatherLow: Float){
 
     fun getHeightVisuals(height: Int?) : Int {
         return when (height) {
